last commit

+++ attendance_management/src/main/java/com/attendance/attendance_management/repository/AttendanceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AttendanceRepository extends JpaRepository<AttendanceInfo, Integer> {
+public interface AttendanceRepository extends JpaRepository<AttendanceInfo, Long> {
 
 }
